requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0